Minor changes to styling, also moving resolvers to main component folder
