+++ src/InterviewQ/components/Review/ReviewPage.scss
@@ -1,7 +1,7 @@
 @import '../../../global/styles/variables';
-$width: 80%;
+$width: 50%;
 .review-page {
-    margin: auto;
+    margin: 6% auto;
     text-align: left;
     width: $width;
     h2 {
