+++ src/InterviewQ/components/Review/subs/ReviewForm.scss
@@ -50,7 +50,7 @@
       width: 15%;
       margin: 0 5%;
       border-radius: 8px;
-      padding: 3% 5%;
+      padding: 2.8% 5%;
     }
     .cancel {
       color:gray;
